+++ packages/webpack/src/plugin.js
@@ -176,19 +176,24 @@ class LavaMoatPlugin {
       STORE.options.readableResourceIds = compiler.options.mode !== 'production'
     }
 
-    let FORCED_CONCATENATEMODULES_OFF = false
+    /** @type {string[]} */
+    const FORCED_CONFIG = []
     if (compiler.options.optimization.concatenateModules) {
-      FORCED_CONCATENATEMODULES_OFF = true
+      FORCED_CONFIG.push('concatenateModules=true->false')
+    }
+    if (compiler.options.optimization.sideEffects) {
+      FORCED_CONFIG.push('sideEffects=true->false')
     }
     // Concatenation won't work with wrapped modules. Have to disable it.
     compiler.options.optimization.concatenateModules = false
+    // package policy enforcement won't work if webpack takes shortcuts
+    compiler.options.optimization.sideEffects = false
 
     // TODO: Research. If we fiddle a little with how we wrap the module, it might be possible to get inlining to work eventually by adding a closure that returns the module namespace. I just don't want to get into the compatibility of it all yet.
     // TODO: explore how these settings affect the Compartment wrapping etc.
     // compiler.options.optimization.mangleExports = false;
     // compiler.options.optimization.usedExports = false;
     // compiler.options.optimization.providedExports = false;
-    // compiler.options.optimization.sideEffects = false;
     Object.freeze(STORE.options)
 
     // =======================================
@@ -246,12 +251,12 @@ class LavaMoatPlugin {
         }
 
         if (
-          FORCED_CONCATENATEMODULES_OFF ||
-          STORE.options.diagnosticsVerbosity > 0
+          FORCED_CONFIG.length > 0
         ) {
           STORE.mainCompilationWarnings.push(
             new WebpackError(
-              'LavaMoatPlugin: Concatenation of modules disabled - not compatible with LavaMoat wrapped modules.'
+              'LavaMoatPlugin: Following options had to be overriden for security: ' +
+                FORCED_CONFIG.join(', ')
             )
           )
         }